pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "royalaudit-digitizer"
version = "2.0.0"
description = "Enterprise-grade invoice extraction system using YOLOv5 for British financial documents"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Ugur Tuna", email = "ugurtuna@example.com"}
]
maintainers = [
    {name = "Ugur Tuna", email = "ugurtuna@example.com"}
]
keywords = [
    "computer-vision",
    "yolov5",
    "document-ai",
    "invoice-extraction",
    "object-detection",
    "pytorch",
    "ocr",
    "mlops",
    "deep-learning"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: GPU :: NVIDIA CUDA",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Typing :: Typed"
]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.0.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.25.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "httpx>=0.26.0",
    "structlog>=24.0.0",
    "prometheus-client>=0.19.0",
    "tenacity>=8.2.0",
    "orjson>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-benchmark>=4.0.0",
    "hypothesis>=6.96.0",
    "mypy>=1.8.0",
    "ruff>=0.1.14",
    "black>=24.1.0",
    "pre-commit>=3.6.0",
    "bandit>=1.7.7",
    "safety>=2.3.5",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]
streamlit = [
    "streamlit>=1.30.0",
    "watchdog>=4.0.0",
]
all = [
    "royalaudit-digitizer[dev,docs,streamlit]",
]

[project.urls]
Homepage = "https://github.com/dsugurtuna/british-invoice-digitization"
Documentation = "https://github.com/dsugurtuna/british-invoice-digitization#readme"
Repository = "https://github.com/dsugurtuna/british-invoice-digitization.git"
Issues = "https://github.com/dsugurtuna/british-invoice-digitization/issues"
Changelog = "https://github.com/dsugurtuna/british-invoice-digitization/blob/main/CHANGELOG.md"

[project.scripts]
royalaudit = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["py.typed", "*.yaml", "*.yml", "*.json"]

# ============================================================================
# BLACK - Code Formatter
# ============================================================================
[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
)/
'''

# ============================================================================
# RUFF - Linter (replaces flake8, isort, pylint)
# ============================================================================
[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pypackages__",
    "build",
    "dist",
    "migrations",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # pylint
    "RUF",    # ruff-specific rules
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit (security)
    "DTZ",    # flake8-datetimez
    "LOG",    # flake8-logging
    "G",      # flake8-logging-format
    "TRY",    # tryceratops
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "PLR0913", # too many arguments
    "TRY003", # long exception messages
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG", "PLR2004"]
"notebooks/**/*.py" = ["E402", "F401"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-single-line = false
lines-between-types = 1

# ============================================================================
# MYPY - Static Type Checker
# ============================================================================
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true
exclude = [
    "venv",
    "build",
    "dist",
    "migrations",
]

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "torchvision.*",
    "cv2.*",
    "ultralytics.*",
    "PIL.*",
    "streamlit.*",
]
ignore_missing_imports = true

# ============================================================================
# PYTEST - Testing Framework
# ============================================================================
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "-p", "no:warnings",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=xml:coverage.xml",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=80",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "gpu: marks tests that require GPU",
]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
parallel = true
omit = [
    "src/__main__.py",
    "tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]

# ============================================================================
# BANDIT - Security Linter
# ============================================================================
[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101"]  # Skip assert warnings in tests
